feat: add stub scripts, s4c-scripts
