+++ scripts-legacy/s4c_cli.py
