+++ scripts/s4c
@@ -1,0 +1,1 @@
+../s4c-scripts/s4c
